requirements.txt
pyaudio==0.2.14
openai==1.54.0
openai-whisper==20231117
pynput==1.7.7
keyboard==0.13.5
pywin32==306
sounddevice==0.4.7
numpy==1.26.4
scipy==1.13.1
torch==2.3.0
tiktoken==0.7.0
python-dotenv==1.0.1
requests==2.32.3
